move BasisState to qre.ops
